pyproject.toml
[project]
name = "insight-hub"
version = "0.1.0"
description = "plafaforma de insights economicos"
authors = [
    {name = "leandrosfreitas",email = "leandrosanfreitas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "fastapi (>=0.128.5,<0.129.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "bcrypt (==4.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
